pyproject.toml
[project]
name = "traffic-sandbox"
version = "0.1.0"
description = "A lightweight, Docker-based traffic simulation sandbox for autonomous driving testing"
requires-python = ">=3.10"
dependencies = [
    "loguru",
    "numpy",
    "shapely",
    "networkx",
    "pyzmq",
    "msgpack",
    "msgpack-numpy",
    "scipy",
    "omegaconf",
    "hydra-core",
    "tinyrpc==1.1.7",
    "apollo-modules==7.0.2",
    "rtree",
    "python-socketio==5.11.4",
    "python-engineio==4.10.1",
    "flask==3.0.3",
    "flask-socketio==5.4.1",
    "eventlet",
]

[project.scripts]
traffic-sandbox = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
